refactor(push): extract push-open handling from Klaviyo.kt (MAGE-768 follow-up) (#500)
* refactor(push): extract push-open handling from Klaviyo into KlaviyoPushOpenHandler


* test(push): move handlePush tests into KlaviyoPushOpenHandlerTest


* test(push): address CodeRabbit feedback on KlaviyoPushOpenHandlerTest

- Add coverage for the notification-dismissal branch
- Replace !! with requireNotNull for _k extra lookups
- Assert warning log level rather than exact message text
- Drop redundant Registry.unregister<Config>() in cleanup

-Claude

---------
