+++ sdk/analytics/src/test/java/com/klaviyo/analytics/KlaviyoTest.kt
@@ -18,13 +18,11 @@ import com.klaviyo.analytics.state.KlaviyoState
 import com.klaviyo.analytics.state.ProfileEventObserver
 import com.klaviyo.analytics.state.State
 import com.klaviyo.analytics.state.StateSideEffects
-import com.klaviyo.core.Constants
 import com.klaviyo.core.DeviceProperties
 import com.klaviyo.core.Registry
 import com.klaviyo.core.config.Config
 import com.klaviyo.core.config.MissingAPIKey
 import com.klaviyo.fixtures.BaseTest
-import com.klaviyo.fixtures.MockIntent
 import com.klaviyo.fixtures.mockDeviceProperties
 import com.klaviyo.fixtures.unmockDeviceProperties
 import io.mockk.coEvery
@@ -57,27 +55,6 @@ import org.junit.Test
 internal class KlaviyoTest : BaseTest() {
 
     companion object {
-        val stubIntentExtras = mapOf(
-            "com.klaviyo.body" to "Message body",
-            "com.klaviyo._k" to """{
-              "Push Platform": "android",
-              "$\flow": "",
-              "$\message": "01GK4P5W6AV4V3APTJ727JKSKQ",
-              "$\variation": "",
-              "Message Name": "check_push_pipeline",
-              "Message Type": "campaign",
-              "c": "6U7nPA",
-              "cr": "31698553996657051350694345805149781",
-              "m": "01GK4P5W6AV4V3APTJ727JKSKQ",
-              "t": "1671205224",
-              "timestamp": "2022-12-16T15:40:24.049427+00:00",
-              "x": "manual"
-            }"""
-        )
-
-        fun mockIntent(payload: Map<String, String>, uri: Uri? = null) =
-            MockIntent.mockIntentWith(payload, uri).intent
-
         private const val TRACKING_URL = "https://trk.klaviyo.com/u/slug"
 
         private const val DESTINATION_URL = "https://www.klaviyo.com/some/path?query=param"
@@ -610,314 +587,6 @@ internal class KlaviyoTest : BaseTest() {
     fun `Fetches push token from persistent store`() {
         spyDataStore.store("push_token", PUSH_TOKEN)
         assertEquals(Klaviyo.getPushToken(), PUSH_TOKEN)
-    }
-
-    private fun verifyOpenedPushEventEnqueued() = verify(exactly = 1) {
-        mockApiClient.enqueueEvent(
-            match { event -> event.metric == EventMetric.OPENED_PUSH },
-            any()
-        )
-    }
-
-    private fun captureOpenedPushEvent() = slot<Event>().also {
-        every { mockApiClient.enqueueEvent(capture(it), any()) } returns mockk(relaxed = true)
-    }
-
-    private fun setupDeepLinkHandler(): Pair<() -> Uri?, DeepLinkHandler> {
-        var capturedUri: Uri? = null
-        Klaviyo.registerDeepLinkHandler { uri: Uri -> capturedUri = uri }
-        return { capturedUri } to Registry.get<DeepLinkHandler>()
-    }
-
-    @Test
-    fun `Non-klaviyo or null intents are ignored`() {
-        // doesn't have _k, klaviyo tracking params
-        Klaviyo.handlePush(mockIntent(mapOf("com.other.package.message" to "3rd party push")))
-        Klaviyo.handlePush(null)
-
-        verify(inverse = true) { mockApiClient.enqueueEvent(any(), any()) }
-    }
-
-    @Test
-    fun `handlePush enqueues opened_push event for klaviyo push intent`() {
-        Klaviyo.handlePush(mockIntent(stubIntentExtras))
-        verifyOpenedPushEventEnqueued()
-    }
-
-    @Test
-    fun `handlePush includes klaviyo extras and push token in opened_push event`() {
-        val eventSlot = captureOpenedPushEvent()
-        Registry.get<State>().pushToken = PUSH_TOKEN
-
-        Klaviyo.handlePush(mockIntent(stubIntentExtras))
-
-        assertTrue(eventSlot.isCaptured)
-        val capturedEvent = eventSlot.captured
-        assertEquals(EventMetric.OPENED_PUSH, capturedEvent.metric)
-
-        // Verify that klaviyo extras are included (with com.klaviyo. prefix removed)
-        assertNotNull(capturedEvent[EventKey.CUSTOM("body")])
-        assertNotNull(capturedEvent[EventKey.CUSTOM("_k")])
-
-        // Verify push token was in the event
-        assertEquals(PUSH_TOKEN, eventSlot.captured[EventKey.PUSH_TOKEN])
-    }
-
-    @Test
-    fun `handlePush invokes DeepLinkHandler when registered and intent has URI data`() {
-        val (getCapturedUri) = setupDeepLinkHandler()
-        val testUri = mockk<Uri>()
-
-        Klaviyo.handlePush(mockIntent(stubIntentExtras, testUri))
-
-        assertEquals(testUri, getCapturedUri())
-    }
-
-    @Test
-    fun `handlePush does not invoke DeepLinkHandler when not registered`() {
-        val testUri = mockk<Uri>()
-
-        Klaviyo.handlePush(mockIntent(stubIntentExtras, testUri))
-
-        verifyOpenedPushEventEnqueued()
-        verify(inverse = true) { DeepLinking.handleDeepLink(testUri) }
-    }
-
-    @Test
-    fun `handlePush does not invoke DeepLinkHandler when intent has no URI data`() {
-        val (getCapturedUri) = setupDeepLinkHandler()
-
-        Klaviyo.handlePush(null)
-        Klaviyo.handlePush(mockIntent(stubIntentExtras))
-
-        assertEquals(null, getCapturedUri())
-        verifyOpenedPushEventEnqueued()
-    }
-
-    @Test
-    fun `handlePush does not invoke deep link handler for non-klaviyo push intents`() {
-        val (getCapturedUri) = setupDeepLinkHandler()
-        val testUri = mockk<Uri>()
-
-        Klaviyo.handlePush(mockIntent(mapOf("some.other.extra" to "value"), testUri))
-
-        assertEquals(null, getCapturedUri())
-        verify(inverse = true) { mockApiClient.enqueueEvent(any(), any()) }
-    }
-
-    // --- Push-open dedup guard (keyed on `_k.tm`, else the generated NOTIFICATION_UID_EXTRA) ---
-    // Note: Klaviyo is an object, so its in-memory dedup set persists across tests in this class.
-    // Each dedup test uses a distinct delivery ID to stay independent of execution order.
-
-    private fun deliveryExtras(deliveryId: String) = mapOf(
-        "com.klaviyo.body" to "Message body",
-        "com.klaviyo._k" to """{"m":"01GK4P5W6AV4V3APTJ727JKSKQ","tm":"$deliveryId"}"""
-    )
-
-    private fun deliveryIntent(deliveryId: String, uri: Uri? = null): Intent =
-        mockIntent(deliveryExtras(deliveryId), uri)
-
-    /** A Klaviyo intent whose `_k` payload omits `tm`, relying on the generated uid as the key. */
-    private fun tmLessIntent(notificationUid: String): Intent =
-        mockIntent(
-            mapOf(
-                "com.klaviyo._k" to """{"m":"01GK4P5W6AV4V3APTJ727JKSKQ"}""",
-                Constants.NOTIFICATION_UID_EXTRA to notificationUid
-            )
-        )
-
-    @Test
-    fun `handlePush tracks exactly one opened_push when a delivery is handled twice`() {
-        // The trampoline tracks first, then forwards the same intent (same tm) to a host that still
-        // calls handlePush manually; only one open, dismissal, and deep link result.
-        val (getCapturedUri) = setupDeepLinkHandler()
-        val testUri = mockk<Uri>()
-
-        Klaviyo.handlePush(deliveryIntent("dedup-twice-distinct-intents", testUri))
-        Klaviyo.handlePush(deliveryIntent("dedup-twice-distinct-intents", testUri))
-
-        verifyOpenedPushEventEnqueued()
-        assertEquals(testUri, getCapturedUri())
-        verify(exactly = 1) { DeepLinking.handleDeepLink(testUri) }
-    }
-
-    @Test
-    fun `handlePush short-circuits a repeat call with the same intent`() {
-        val intent = deliveryIntent("dedup-same-intent-object")
-
-        Klaviyo.handlePush(intent)
-        Klaviyo.handlePush(intent)
-
-        verifyOpenedPushEventEnqueued()
-    }
-
-    @Test
-    fun `handlePush tracks distinct deliveries independently`() {
-        Klaviyo.handlePush(deliveryIntent("dedup-distinct-A"))
-        Klaviyo.handlePush(deliveryIntent("dedup-distinct-B"))
-
-        verify(exactly = 2) {
-            mockApiClient.enqueueEvent(match { it.metric == EventMetric.OPENED_PUSH }, any())
-        }
-    }
-
-    @Test
-    fun `handlePush dedupes a tm-less delivery via the generated notification uid`() {
-        Klaviyo.handlePush(tmLessIntent("uid-fallback-same"))
-        Klaviyo.handlePush(tmLessIntent("uid-fallback-same"))
-
-        verifyOpenedPushEventEnqueued()
-    }
-
-    @Test
-    fun `handlePush tracks distinct tm-less notifications independently`() {
-        // Distinct notifications get distinct generated uids, so they must not collide.
-        Klaviyo.handlePush(tmLessIntent("uid-fallback-A"))
-        Klaviyo.handlePush(tmLessIntent("uid-fallback-B"))
-
-        verify(exactly = 2) {
-            mockApiClient.enqueueEvent(match { it.metric == EventMetric.OPENED_PUSH }, any())
-        }
-    }
-
-    @Test
-    fun `handlePush falls back to the generated uid when the _k payload is not valid JSON`() {
-        // A non-JSON _k must fail tm parsing gracefully and fall back to the generated uid.
-        val makeIntent = {
-            mockIntent(
-                mapOf(
-                    "com.klaviyo._k" to "not-json",
-                    Constants.NOTIFICATION_UID_EXTRA to "uid-malformed-k"
-                )
-            )
-        }
-
-        Klaviyo.handlePush(makeIntent())
-        Klaviyo.handlePush(makeIntent())
-
-        verifyOpenedPushEventEnqueued()
-    }
-
-    @Test
-    fun `handlePush does not dedupe when no delivery id is available`() {
-        // No `tm` and no generated uid: with no key to match on, each call tracks. We never fabricate
-        // a key from the shared `_k` metadata, which would collapse distinct deliveries.
-        val noKey = mapOf("com.klaviyo._k" to """{"m":"01GK4P5W6AV4V3APTJ727JKSKQ"}""")
-
-        Klaviyo.handlePush(mockIntent(noKey))
-        Klaviyo.handlePush(mockIntent(noKey))
-
-        verify(exactly = 2) {
-            mockApiClient.enqueueEvent(match { it.metric == EventMetric.OPENED_PUSH }, any())
-        }
-    }
-
-    @Test
-    fun `handlePush continues deep link handling even if opened_push processing fails`() {
-        val (getCapturedUri) = setupDeepLinkHandler()
-        every { mockApiClient.enqueueEvent(any(), any()) } throws MissingAPIKey()
-        val testUri = mockk<Uri>()
-
-        Klaviyo.handlePush(mockIntent(stubIntentExtras, testUri))
-
-        // Deep link handler should still be invoked despite the API error
-        assertEquals(testUri, getCapturedUri())
-    }
-
-    @Test
-    fun `handlePush decodes valid key_value_pairs JSON into a map`() {
-        val eventSlot = captureOpenedPushEvent()
-        val keyValuePairsJson = """{"custom_key_1":"value1","custom_key_2":"value2"}"""
-        val extrasWithKeyValuePairs = mapOf(
-            "com.klaviyo._k" to stubIntentExtras["com.klaviyo._k"]!!,
-            "com.klaviyo.key_value_pairs" to keyValuePairsJson
-        )
-
-        Klaviyo.handlePush(mockIntent(extrasWithKeyValuePairs))
-
-        assertTrue(eventSlot.isCaptured)
-        val capturedEvent = eventSlot.captured
-        val keyValuePairs = capturedEvent[EventKey.CUSTOM("key_value_pairs")]
-
-        // Verify that the value is a map, not a string
-        assertTrue(keyValuePairs is Map<*, *>)
-        val map = keyValuePairs as Map<*, *>
-        assertEquals("value1", map["custom_key_1"])
-        assertEquals("value2", map["custom_key_2"])
-        assertEquals(2, map.size)
-    }
-
-    @Test
-    fun `handlePush falls back to raw string when key_value_pairs JSON is invalid`() {
-        val eventSlot = captureOpenedPushEvent()
-        val invalidJson = """{"invalid": "json"""
-        val extrasWithInvalidKeyValuePairs = mapOf(
-            "com.klaviyo._k" to stubIntentExtras["com.klaviyo._k"]!!,
-            "com.klaviyo.key_value_pairs" to invalidJson
-        )
-
-        Klaviyo.handlePush(mockIntent(extrasWithInvalidKeyValuePairs))
-
-        assertTrue(eventSlot.isCaptured)
-        val capturedEvent = eventSlot.captured
-        val keyValuePairs = capturedEvent[EventKey.CUSTOM("key_value_pairs")]
-
-        // Verify that the value falls back to the raw string
-        assertTrue(keyValuePairs is String)
-        assertEquals(invalidJson, keyValuePairs)
-
-        // Verify warning was logged
-        verify {
-            spyLog.warning(
-                match { it.contains("Failed to parse key_value_pairs JSON") },
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `handlePush decodes empty key_value_pairs JSON into empty map`() {
-        val eventSlot = captureOpenedPushEvent()
-        val emptyJson = """{}"""
-        val extrasWithEmptyKeyValuePairs = mapOf(
-            "com.klaviyo._k" to stubIntentExtras["com.klaviyo._k"]!!,
-            "com.klaviyo.key_value_pairs" to emptyJson
-        )
-
-        Klaviyo.handlePush(mockIntent(extrasWithEmptyKeyValuePairs))
-
-        assertTrue(eventSlot.isCaptured)
-        val capturedEvent = eventSlot.captured
-        val keyValuePairs = capturedEvent[EventKey.CUSTOM("key_value_pairs")]
-
-        // Verify that the value is an empty map
-        assertTrue(keyValuePairs is Map<*, *>)
-        val map = keyValuePairs as Map<*, *>
-        assertTrue(map.isEmpty())
-    }
-
-    @Test
-    fun `handlePush still decodes other klaviyo extras as strings`() {
-        val eventSlot = captureOpenedPushEvent()
-        val extrasWithMultipleFields = mapOf(
-            "com.klaviyo._k" to stubIntentExtras["com.klaviyo._k"]!!,
-            "com.klaviyo.body" to "Test message",
-            "com.klaviyo.title" to "Test title"
-        )
-
-        Klaviyo.handlePush(mockIntent(extrasWithMultipleFields))
-
-        assertTrue(eventSlot.isCaptured)
-        val capturedEvent = eventSlot.captured
-
-        // Verify other fields are still strings
-        val body = capturedEvent[EventKey.CUSTOM("body")]
-        val title = capturedEvent[EventKey.CUSTOM("title")]
-
-        assertTrue(body is String)
-        assertTrue(title is String)
-        assertEquals("Test message", body)
-        assertEquals("Test title", title)
     }
 
     @Test
